convert from flatten triangular to matrix representation before doing matrix multiplications
